ai-learning-flask-project (2)/ai-learning-flask/requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.3
numpy==1.26.4
scikit-learn==1.5.0
reportlab==4.2.2
python-dotenv==1.0.1
Pillow==10.3.0
firebase-admin==6.5.0
google-genai>=1.0.0
